Insert and delete ChiTietDichVuThuePhong

+++ src/QLKS/DTO/ChiTietTienThuePhong.cs
@@ -14,11 +14,11 @@ namespace QLKS.DTO
 
         public string LoaiTienPhong { get; set; }
 
-        public float? Gia { get; set; }
+        public double? Gia { get; set; }
 
         public int SoLuong { get; set; }
 
-        public float? ThanhTien { get; set; }
+        public double? ThanhTien { get; set; }
 
         public DateTime? NgayTao { get; set; }
